+++ .flaskenv
@@ -1,3 +1,2 @@
 GOOGLE_CLIENT_ID = '394376551777-aal6bi33dtbu3t1hu09nm240mju398pf.apps.googleusercontent.com'
 GOOGLE_CLIENT_SECRET='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'
-secret_key='mycloset'
