+++ templates/index.html
@@ -39,9 +39,9 @@
 <body>
     <h1>Hello World!</h1>
     {% if logged %}
-        <a href="/logout">로그아웃</a>
+        <a href="/login/logout">로그아웃</a>
     {% else %}
-        <a href="/login_page">로그인페이지</a>
+        <a href="/login/login_page">로그인페이지</a>
     {% endif %}
 </body>
