+++ templates/login.html
@@ -34,13 +34,13 @@
 </head>
 <body>
     {% if not google_data %}
-        <form name="f_login" method="post" enctype="multipart/form-data" action="/login">
+        <form name="f_login" method="post" enctype="multipart/form-data" action="/login/login">
             ID : <input type="text" name="user_id" id="user_id">
             PW : <input type="password" name="user_pw" id="user_pw">
 
             <input type="submit" value="로그인하기">
         </form>
-            <a href="/google_login"><img src="../static/google_login.png"></a>
+            <a href="/login/google_login"><img src="../static/google_login.png"></a>
     {% else %}
         안녕하세요 {{google_data.given_name}}님.
     {% endif %}
